requirements.txt
git+https://git@github.com/ping/instagram_private_api.git@1.6.0
typer==0.2.1
pytest==5.3.4
pytest-cov==2.8.1
pydantic==1.4
python-dotenv==0.12.0
loguru==0.4.1
